Part 1 of the dap4 refactor. 1. Move all cdm dependent code into dap4/d4cdm 2. Move all  code in d4client into dap4/d4shared    (will probably undo partly in part2 or later) 3. Move all code in d4tswar into dap4/d4ts 4. Misc. other mods to support above 5. Modify various gradle files to support above.
