+++ dap4/build.gradle
@@ -6,10 +6,7 @@ ext.title = "Data Access Protocol (DAP) version 4.0"  // Will be inherited by su
 // common:
 //	d4core		nothing
 // 	d4shared	d4core httpservices
-// 	d4cdmshared	d4shared cdm
-// client:
-// 	d4client	d4shared
-// 	d4cdmclient	d4cdmshared
+// 	d4cdm		d4shared cdm
 // server:
 // 	d4servletshared	d4cdmshared netcdf4
 // 	d4ts		d4servletshared
