+++ dap4/d4cdm/build.gradle
@@ -1,7 +1,6 @@
 dependencies {
-    compile project(':dap4:d4core')
     compile project(':dap4:d4shared')
-
+    compile project(':cdm')
+    compile project(':netcdf4')
     provided libraries["javax.servlet-api"]
-    compile libraries["slf4j-api"]
 }
