+++ dap4/d4cdm/src/main/java/dap4/cdm/CDMArray.java
