+++ dap4/d4shared/build.gradle
@@ -2,6 +2,9 @@ dependencies {
     compile project(':dap4:d4core')
     compile project(':httpservices')
 
+    compile libraries["jcommander"]
+    compile libraries["slf4j-api"]
+    runtime libraries["slf4j-jdk14"]
     compile libraries["httpclient"]
     compile libraries["httpcore"]
 }
