+++ dap4/d4ts/src/main/java/dap4/d4ts/D4TSServlet.java
@@ -37,7 +37,6 @@ public class D4TSServlet extends DapController
             // Register known DSP classes: order is important.
             // Only used in server
             registerDSP(SynDSP.class, true);
-            registerDSP(CDMDSP.class, true);
         }
 
     }
