+++ gradle/fatJars.gradle
@@ -7,7 +7,7 @@ configurations {
 }
 
 dependencies {
-    dap4 project(':dap4:d4client')
+    dap4 project(':dap4:d4shared')
 
     ncIdv project(':cdm')
     ncIdv project(':clcommon')
@@ -52,9 +52,9 @@ fatJarTasks << tasks.create(name: 'buildDap4', type: ShadowJar) {
     configurations = [project.configurations.dap4]
 
     doFirst {
-        // This script is evaluated before :dap4:d4client, so we must wait until execution time to get its properties.
+        // This script is evaluated before :dap4:d4shared, so we must wait until execution time to get its properties.
         // This does not affect the incremental build. The main disadvantage is that failure (if any) is delayed.
-        manifest.attributes project(':dap4:d4client').tasks.jar.manifest.attributes
+        manifest.attributes project(':dap4:d4shared').tasks.jar.manifest.attributes
     }
 }
