+++ settings.gradle
@@ -6,15 +6,12 @@ include 'bufr'
 include 'cdm'
 include 'cdm-test'
 include 'clcommon'
-include 'dap4:d4cdmclient'
-include 'dap4:d4cdmshared'
-include 'dap4:d4client'
-include 'dap4:d4core'
+include 'dap4:d4cdm'
 include 'dap4:d4servletshared'
 include 'dap4:d4shared'
+include 'dap4:d4core'
 include 'dap4:d4tests'
 include 'dap4:d4ts'
-include 'dap4:d4tswar'
 include 'grib'
 include 'httpservices'
 include 'it'
