+++ tds/build.gradle
@@ -17,7 +17,7 @@ dependencies {
     compile project(":waterml")
 
     // DAP4 Dependencies (technically forward)
-    compile project(":dap4:d4core")
+    compile project(":dap4:d4cdm")
     compile project(":dap4:d4servletshared")
 
     // Server stuff
